Changes Dec 28th
Added Comment controller which handles all the requests to add a comment to a post.

Added Comment model and Routes as well.

Modified userController, so once the user uploads a new image, the previous one is deleted

Modified UserRoutes and added changepwd so a user can change password,

modified PostController to send Comments information to the frontend